# learning-platform/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.TaskResponse:
    properties:
      answerType:
        type: string
      authorId:
        type: string
      bodyMd:
        type: string
      correctAnswer:
        type: string
      createdAt:
        type: string
      difficulty:
        type: string
      id:
        type: string
      imageUrl:
        type: string
      officialSolution:
        type: string
      status:
        type: string
      title:
        type: string
      topicId:
        type: string
      updatedAt:
        type: string
    type: object
...
  response.ErrorMessage:
    properties:
      message:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorMessage'
      success:
        example: false
        type: boolean
    type: object
  response.SuccessWrapper:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for Learning Platform — users, auth, tasks, topics
  title: Learning Platform API
  version: "1.0"
paths:
...
  /tasks:
    get:
      description: Returns list of all published tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.TaskResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get all published tasks
      tags:
      - tasks
    post:
      consumes:
      - multipart/form-data
      description: Create a task (multipart form with optional image)
      parameters:
      - description: Title
        in: formData
        name: title
        required: true
        type: string
      - description: Markdown body
        in: formData
        name: body_md
        required: true
        type: string
      - description: Difficulty
        in: formData
        name: difficulty
        required: true
        type: string
      - description: Task status
        in: formData
        name: status
        required: true
        type: string
      - description: Topic ID
        in: formData
        name: topicId
        required: true
        type: string
      - description: Solution
        in: formData
        name: officialSolution
        required: true
        type: string
      - description: Correct answer
        in: formData
        name: correctAnswer
        required: true
        type: string
      - description: Answer type
        in: formData
        name: answerType
        required: true
        type: string
      - description: Task image
        in: formData
        name: imageUrl
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Remove task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a task
      tags:
      - tasks
    get:
      description: Returns a single task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get task by ID
      tags:
      - tasks
    put:
      consumes:
      - multipart/form-data
      description: Update task metadata + (optional) image
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Title
        in: formData
        name: title
        required: true
        type: string
      - description: Body
        in: formData
        name: body_md
        required: true
        type: string
      - description: Difficulty
        in: formData
        name: difficulty
        required: true
        type: string
      - description: Status
        in: formData
        name: status
        required: true
        type: string
      - description: Topic ID
        in: formData
        name: topicId
        required: true
        type: string
      - description: Solution
        in: formData
        name: officialSolution
        required: true
        type: string
      - description: Correct answer
        in: formData
        name: correctAnswer
        required: true
        type: string
      - description: Answer type
        in: formData
        name: answerType
        required: true
        type: string
      - description: Task image
        in: formData
        name: imageUrl
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - tasks
  /tasks/{id}/publish:
    patch:
      description: Changes task status to PUBLISHED
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Publish a task
      tags:
      - tasks
  /tasks/drafts:
    get:
      description: Returns tasks with status DRAFT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.TaskResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all draft tasks
      tags:
      - tasks
  /tasks/my:
    get:
      description: Returns all tasks where author == current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.TaskResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get tasks created by current user
      tags:
      - tasks
...
  /topics/{topicId}/tasks:
    get:
      description: Returns all tasks belonging to topic
      parameters:
      - description: Topic ID
        in: path
        name: topicId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessWrapper'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.TaskResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get tasks by topic
      tags:
      - tasks
...
swagger: "2.0"